mvoe functions to general auth
