+++ src/frontend/src/pages/Auth/MFASetup.tsx
@@ -1,11 +1,8 @@
 import { Trans, t } from '@lingui/macro';
 import { Button } from '@mantine/core';
-import { showNotification } from '@mantine/notifications';
 import { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
-import { ApiEndpoints } from '../../enums/ApiEndpoints';
-import { authApi, followRedirect } from '../../functions/auth';
-import { apiUrl } from '../../states/ApiState';
+import { getTotpSecret, handleVerifyTotp } from '../../functions/auth';
 import { QrRegistrationForm } from '../Index/Settings/AccountSettings/QrRegistrationForm';
 import { Wrapper } from './LoginLayoutComponent';
 
@@ -16,28 +13,9 @@ export default function MfaSetup() {
   const [totpQr, setTotpQr] = useState<{ totp_url: string; secret: string }>();
   const [value, setValue] = useState('');
 
-  const registerTotp = async () => {
-    await authApi(apiUrl(ApiEndpoints.auth_totp), undefined, 'get').catch(
-      (err) => {
-        if (err.status == 404 && err.response.data.meta.secret) {
-          setTotpQr(err.response.data.meta);
-        } else {
-          const msg = err.response.data.errors[0].message;
-          showNotification({
-            title: t`Failed to set up MFA`,
-            message: msg,
-            color: 'red'
-          });
-        }
-      }
-    );
-  };
-
   useEffect(() => {
-    if (!totpQr) {
-      registerTotp();
-    }
-  }, [totpQr]);
+    getTotpSecret(setTotpQr);
+  }, []);
 
   return (
     <Wrapper titleText={t`MFA Setup Required`} logOff>
@@ -49,13 +27,7 @@ export default function MfaSetup() {
       />
       <Button
         disabled={!value}
-        onClick={() => {
-          authApi(apiUrl(ApiEndpoints.auth_totp), undefined, 'post', {
-            code: value
-          }).then(() => {
-            followRedirect(navigate, location?.state);
-          });
-        }}
+        onClick={handleVerifyTotp(value, navigate, location)}
       >
         <Trans>Add TOTP</Trans>
       </Button>
