+++ src/frontend/src/pages/Auth/ResetPassword.tsx
@@ -5,80 +5,26 @@ import { notifications } from '@mantine/notifications';
 import { useEffect } from 'react';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 
-import { api } from '../../App';
-import { ApiEndpoints } from '../../enums/ApiEndpoints';
-import { apiUrl } from '../../states/ApiState';
+import { handlePasswordReset } from '../../functions/auth';
 import { Wrapper } from './LoginLayoutComponent';
 
 export default function ResetPassword() {
   const simpleForm = useForm({ initialValues: { password: '' } });
   const [searchParams] = useSearchParams();
   const navigate = useNavigate();
-
   const key = searchParams.get('key');
 
-  function invalidKey() {
-    notifications.show({
-      title: t`Key invalid`,
-      message: t`You need to provide a valid key to set a new password. Check your inbox for a reset link.`,
-      color: 'red'
-    });
-    navigate('/login');
-  }
-
-  function success() {
-    notifications.show({
-      title: t`Password set`,
-      message: t`The password was set successfully. You can now login with your new password`,
-      color: 'green',
-      autoClose: false
-    });
-    navigate('/login');
-  }
-
-  function passwordError(values: any) {
-    notifications.show({
-      title: t`Reset failed`,
-      message: values?.errors.map((e: any) => e.message).join('\n'),
-      color: 'red'
-    });
-  }
-
+  // make sure we have a key
   useEffect(() => {
-    // make sure we have a key
     if (!key) {
-      invalidKey();
+      notifications.show({
+        title: t`Key invalid`,
+        message: t`You need to provide a valid key to set a new password. Check your inbox for a reset link.`,
+        color: 'red'
+      });
+      navigate('/login');
     }
   }, [key]);
-
-  function handleSet() {
-    // Set password with call to backend
-    api
-      .post(
-        apiUrl(ApiEndpoints.user_reset_set),
-        {
-          key: key,
-          password: simpleForm.values.password
-        },
-        { headers: { Authorization: '' } }
-      )
-      .then((val) => {
-        if (val.status === 200) {
-          success();
-        } else {
-          passwordError(val.data);
-        }
-      })
-      .catch((err) => {
-        if (err.response?.status === 400) {
-          passwordError(err.response.data);
-        } else if (err.response?.status === 401) {
-          success();
-        } else {
-          passwordError(err.response.data);
-        }
-      });
-  }
 
   return (
     <Wrapper titleText={t`Set new password`}>
@@ -88,7 +34,12 @@ export default function ResetPassword() {
         description={t`The desired new password`}
         {...simpleForm.getInputProps('password')}
       />
-      <Button type='submit' onClick={handleSet}>
+      <Button
+        type='submit'
+        onClick={() =>
+          handlePasswordReset(key, simpleForm.values.password, navigate)
+        }
+      >
         <Trans>Send Password</Trans>
       </Button>
     </Wrapper>
