+++ src/frontend/src/pages/Auth/VerifyEmail.tsx
@@ -4,47 +4,28 @@ import { notifications } from '@mantine/notifications';
 import { useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 
-import { api } from '../../App';
-import { ApiEndpoints } from '../../enums/ApiEndpoints';
-import { apiUrl } from '../../states/ApiState';
+import { handleVerifyEmail } from '../../functions/auth';
 import { Wrapper } from './LoginLayoutComponent';
 
 export default function VerifyEmail() {
   const { key } = useParams();
   const navigate = useNavigate();
 
-  function invalidKey() {
-    notifications.show({
-      title: t`Key invalid`,
-      message: t`You need to provide a valid key.`,
-      color: 'red'
-    });
-    navigate('/login');
-  }
-
+  // make sure we have a key
   useEffect(() => {
-    // make sure we have a key
     if (!key) {
-      invalidKey();
+      notifications.show({
+        title: t`Key invalid`,
+        message: t`You need to provide a valid key.`,
+        color: 'red'
+      });
+      navigate('/login');
     }
   }, [key]);
 
-  function handleSet() {
-    // Set password with call to backend
-    api
-      .post(apiUrl(ApiEndpoints.auth_email_verify), {
-        key: key
-      })
-      .then((val) => {
-        if (val.status === 200) {
-          navigate('/login');
-        }
-      });
-  }
-
   return (
     <Wrapper titleText={t`Verify Email`}>
-      <Button type='submit' onClick={handleSet}>
+      <Button type='submit' onClick={() => handleVerifyEmail(key, navigate)}>
         <Trans>Verify</Trans>
       </Button>
     </Wrapper>
